SOC-650 | refactor: move code responsible for username - user id translation
